Changed lib so that users can deal with just clojure data structures
Original usage required that downstream users convert some items to js
datatypes (via clj->js or #js).  This caused an inconsistency when
passing params that contained clojure datastructures.

Updated to a later version of cljs so that namespaced keywords are
preserved when going to/from js.  Also changed the lib to handle all
conversions to/from js so that downstream users only need to pass in
clojure datastructures.